Fix positioning bug when target lost
Poscontrol overrides the target_pos based on the leash length and relies on the actual target pos to be continually fed into it again.
Currently if we lose the target but still wish to land on it - it will stop updating the target, which means the final target pos is the actual  target pos minus the last leash correction modification. Hence wont land exactly on the target